refactor: use new base concrete according its report objects
