+++ src/ReportSystem/Concretes/BorrowPatronHistory.cs
@@ -1,15 +1,25 @@
 using BorrowSystem;
 
-namespace ReportSystem.Concrete;
-public class BorrowPatronHistory : PatronReport
-{
-    public BorrowPatronHistory(BorrowManager BorrowList) : base(BorrowList){}
+namespace Opcion1CanelitaTeam.ReportSystem.Concretes;
 
-    public void PatronBorrowHistoryReport(string PatronName)
+public class BorrowPatronHistory : BaseReport
+{
+    public BorrowPatronHistory(BorrowManager borrowManager) 
+        : base(borrowManager)
+    { }
+    
+    public override void GenerateReport()
     {
-        generateReport();
-        foreach(Borrow borrow in Borrows.GetBorrows()){
-            if(borrow.GetPatron().getName() == PatronName){
+        Console.WriteLine("This is a Patron Borrow Report");
+    }
+
+    public void PatronBorrowHistoryReport(string patronName)
+    {
+        GenerateReport();
+        foreach (Borrow borrow in _borrowManager.GetBorrows())
+        {
+            if (borrow.GetPatron().getName() == patronName)
+            {
                 borrow.BorrowDetails();
             }
         }
